Event argument refactoring, added key events
All events are no passed a single object, which has a "event" property which points to the raw browser event. The object may have addition properties depending on the event. For example, for mouse events, the event object contains position property which is a vector of the mouse event relative to the canvas.

Events are now dispatched in an explicit order (see EVENT_ORDER).