Continued tweaks to event dispatching
Was unecssarily calling dispatch when listening, and then existing out if there were no args.

Changed to where the handler now flags itself if it needs to be dispatched. This is then used to determine whether it should be dispatched on next loop. After it is dispatched, flag is once again set to false.

A little less bug prone, and should be a little more performant.