Setting getContext alpha attribute to false
Setting the context alpha attribute in the getContext to false which should improve rendering performance (and currently, we already draw a background).

This means you cannot currently have a transparent background / canvas in the library. We can add a config option for this in the future if we need it.