Added ability to pause rendering
Added option to pause rendering. When renderEnabled is set to false, and rendering commands will be ignored, and not passed on to canvas.

This can be useful when running something intensive that needs time to build up (such as circle packing). So you can run without overhead of rendering for faster performance, and then render once its ready.