Some refactoring around clearing context background
Refactoring how we clear the background.

Move clear method to context. This puts it in one place, allows us to make some general optimizations, and also helps us when we add SVG support (we now have a defnitive call where we know the screen is being wiped).

Also change to pass around color instances and not hex colors in order to maintain consistency with other APIs.

A context now has an optional backgroundColor attribute. if set, when the canvas is cleared, we will just paint over the existing canvas with the background color. If not color is set, we will just call clearRect.

When clear is called, we also clear the command queue.