Caching RGBA / CSS generation and caching
Renamed color.toRGBA to color.toCSS (to make it clearer that it outputs rgba color in css format).

Also implimented caching for the rgba string within the class, so calling this repeately shouldnt have a major performance impact as the string is only calculated if the underlying rgba values have changed since the css string was last generated.