Changing from style to color (again!)
Changing fillStyle and strokeStyle properties on shapes to fillColor and strokeColor to make it clearer that they take Color object instances, and not other styles such as patterns, hex colors or CSS colors.

Might be a bit confusing switching between the two, but its more descriptive of what it is doing.