Implimented command pattern
Implimented command objects to handle context calls. This will add a little overhead, but allows us to transform calls into other output such as SVG.

Note, for the object pooling, we are not explicitly setting properties to null when returning (for performance reasons), but this can cause an issue if all properties are not intialized correctly when retrieved.